+++ kernel/misc.c
@@ -75,8 +75,8 @@ void show_banner(void)
       ""
    };
 
-   const u32 padding =
-      term_get_cols(get_curr_term()) / 2 - strlen(banner[1]) / 2;
+   const u32 padding = (u32)
+      (term_get_cols(get_curr_term()) / 2 - strlen(banner[1]) / 2);
 
    for (u32 i = 0; i < ARRAY_SIZE(banner); i++)
       console_gfx_replace_chars(banner[i]);
